refactor: extract workspace config from AgentWorkspace
